backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pyodbc==5.1.0
python-dotenv==1.0.1
pydantic==2.7.1

# Azure AI
azure-ai-textanalytics==5.3.0
openai==1.30.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP client
httpx==0.27.0

# Monitoring — Azure Application Insights via OpenTelemetry
azure-monitor-opentelemetry==1.3.0
